+++ src/main/kotlin/goodspace/bllsoneshot/task/mapper/TaskDetailMapper.kt
@@ -1,6 +1,7 @@
 package goodspace.bllsoneshot.task.mapper
 
 import goodspace.bllsoneshot.entity.assignment.Task
+import goodspace.bllsoneshot.mentor.dto.response.MentorTaskFormResponse
 import goodspace.bllsoneshot.task.dto.response.task.TaskDetailResponse
 import org.springframework.stereotype.Component
 
@@ -9,6 +10,17 @@ class TaskDetailMapper(
     private val worksheetMapper: WorksheetMapper,
     private val columnLinkMapper: ColumnLinkMapper
 ) {
+
+    fun mapToForm(task: Task): MentorTaskFormResponse {
+        return MentorTaskFormResponse(
+            subject = task.subject,
+            dates = listOfNotNull(task.date),
+            taskNames = listOf(task.name),
+            goalMinutes = task.goalMinutes,
+            worksheets = task.worksheets.map { worksheetMapper.map(it) },
+            columnLinks = task.columnLinks.map { columnLinkMapper.map(it) }
+        )
+    }
 
     fun map(task: Task): TaskDetailResponse {
         return TaskDetailResponse(
